requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
python-dotenv==1.0.0
pytz==2024.1
supabase>=2.0.0
psycopg2-binary>=2.9.0
httpx>=0.25.2,<0.29

